Cast uid_t and pid_t properly to avoid compiler warnings. See bug #340117
	* daemon/gnome-keyring-daemon-io.c:
	* daemon/gnome-keyring-daemon.c: Cast uid_t and pid_t properly
	to avoid compiler warnings. See bug #340117

svn path=/trunk/; revision=584

+++ daemon/gnome-keyring-daemon-io.c
@@ -392,7 +392,7 @@ gnome_keyring_client_state_machine (GnomeKeyringClient *client)
 		}
 		if (getuid() != uid) {
 			g_warning ("uid mismatch: %u, should be %u\n",
-				   uid, getuid());
+				   (guint)uid, (guint)getuid());
 			gnome_keyring_client_free (client);
 			return;
 		}
